+++ nym-vpn-android/buildSrc/src/main/kotlin/Constants.kt
@@ -1,8 +1,8 @@
 import org.gradle.api.JavaVersion
 
 object Constants {
-	const val VERSION_NAME = "2026.12.3"
-    const val VERSION_CODE = 20261203
+	const val VERSION_NAME = "2026.12.4-beta.1"
+    const val VERSION_CODE = 20261204
     const val TARGET_SDK = 36
     const val COMPILE_SDK = 37
     const val MIN_SDK = 24