+++ nym-vpn-app/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2026.12.4-beta.1] - 2026-09-03
+
 ### Fixed
 
 - Replace lottie animations with manual css animations